Resolved link problem with Tcl_ErrnoMsg and TclRegError.
The problem was that exp_clib.c provides substitute definitions for
some of the Tcl functions when the Tcl library is absent.  (The
library is not used with exp_clib.)  In order for the other functions
in the Expect library (that work with either Tcl/Expect or exp_clib),
they had to be turned into function pointers.